move FileFilterControls component into explore package
