+++ components/HomePage.tsx
@@ -5,10 +5,10 @@ import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
 import Button from 'react-bootstrap/Button';
 import Container from 'react-bootstrap/Container';
 import Jumbotron from 'react-bootstrap/Jumbotron';
-import { AtlasMeta } from '../types';
 import { EntityReport } from '../lib/helpers';
 import { EntityReportByAttribute } from '../lib/entityReportHelpers';
 import Plots from './Plots';
+import { AtlasMeta } from '../packages/data-portal-commons/src/libs/entity';
 
 export interface IHomePropsProps {
     hero_blurb: string;
