+++ packages/data-portal-commons/src/libs/getCaseValues.ts
@@ -1,0 +1,12 @@
+import _ from 'lodash';
+import { Entity } from './entity';
+
+export function getCaseValues(propName: keyof Entity) {
+    return (e: Entity) => {
+        if (e.cases) {
+            return _.uniq(e.cases.map((c) => c[propName] as string));
+        } else {
+            return [e[propName] as string];
+        }
+    };
+}
