+++ packages/data-portal-filter/src/libs/types.ts
@@ -1,4 +1,5 @@
 import { ActionMeta, ActionTypes, OptionTypeBase } from 'react-select';
+import { AttributeMap } from '../../../data-portal-utils/src/libs/types';
 
 export type OptionType = {
     value: string;
@@ -29,4 +30,19 @@ export enum FilterAction {
 export interface FilterActionMeta<OptionType extends OptionTypeBase>
     extends Omit<ActionMeta<OptionType>, 'action'> {
     action: ActionTypes | FilterAction;
+}
+
+export interface IFilterControlsProps<T, Attribute extends string> {
+    setFilter: (actionMeta: any) => void;
+    selectedFiltersByGroupName: ISelectedFiltersByAttrName;
+    selectedFilters: SelectedFilter[];
+    entities: T[];
+    groupsByProperty: { [attrName: string]: { [attrValue: string]: T[] } };
+}
+
+export interface IGenericFilterControlProps<T, Attribute extends string>
+    extends IFilterControlsProps<T, Attribute> {
+    countHeader: string;
+    attributeMap: AttributeMap<T, Attribute>;
+    attributeNames: Attribute[];
 }
