+++ packages/data-portal-utils/src/libs/getDelimitedValues.ts
@@ -1,0 +1,3 @@
+export function getDelimitedValues(text: string, separator: string = ',') {
+    return text.split(separator).map((v) => v.trim());
+}
